Fix format

+++ wasm/test_rust.c
@@ -21,7 +21,7 @@ int main(int argc, char **argv) {
     int i = test_rust();
     printf("In C: test_rust() returned %d\n", i);
     uint8_t *p = &test_rust_buffer[0];
-    printf("In C: test_rust_buffer is 0x%x\n", p);
+    printf("In C: test_rust_buffer is 0x%x\n", (unsigned) p);
     return 0;
 }
